Undo damage done in c6d3d601c96f, which shifted failure to allocate large buffer to pipe_create() from pipe_write().
The need for such allocation by itself is questionable since
absolute majority of users write less than dozen bytes and out
of that majority most write exactly 0.

But that shift created a problem by itself, so that when the
shared resource is exhausted no new pipes can be created
thus preventing any meaningful activity in the system.